SQL Automation/requirements.txt
streamlit
mysql-connector-python
pandas
google-generativeai
python-dotenv